Fix transaction history screen staleness
This commit fixes the staleness in the transaction history screen.
Previously we were not refetching the TRANSACTION_LIST query after
firing the mutations to confirm payments.  Also I have added a swipe down to
refresh on the transaction history screen.  This should not be needed
now that we are properly refetching the queries but it gives the user
confidence that they are up to date.