chore: home screen show usd and stablesats
right now the header on the home screen change when we are going from been log out to be logged in.

this is not a great experience.

plus I think we don't have to make the app generic with variables like hasUsdWallet, as it make the app more complex overall