chore: make deconnection on 401 more rapid (#2943)
I don't think we have had many 401 issue with kratos lately. this has arise in the past where kratos has issue connecting with the associated postgres, kratos was returning 401 (this bug has been reported but not fixed)

the benefits of reducing the retry/initial delay is that when doing some local development, or also when loging out, if there is an issue with the token, there won't be promise that trigger out of the blue and deconnect back the user one minutes after he had log out/log back in. as may be the case today.

also the reason I think that is ok for going to those settings now, is that, with email now here for many months, it is less an issue if a 401 ever appears compare to if we had a phone only login method

Co-authored-by: Nicolas Burtey <nb@galoy.io>